services/paddleocr/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paddleocr==2.7.3
paddlepaddle==3.0.0
pillow==10.2.0
numpy==1.26.3
